+++ demo2.js
@@ -1,5 +1,4 @@
-const DymoServices = require('./src/dymo-services');
-const imageServices = require('./src/image-services');
+const {DymoServices, createImageWithText, convertImageToBitmap} = require('node-dymo-printer');
 
 /**
  * Configure the network printer manually and print "Hello World!" on a 89mm x 36mm label.
@@ -15,16 +14,19 @@ const imageServices = require('./src/image-services');
     };
     const dymoServices = new DymoServices(config);
 
-    const {imageWidth, imageHeight} = dymoServices.DYMO_LABELS[1];
-    const image = await imageServices.createImageWithText(imageWidth, imageHeight, 50, 128, 'Hello World!');
+    // Create landscape image with the dimensions of the label and with the text "Hello World!".
+    const {imageWidth, imageHeight} = DymoServices.DYMO_LABELS[1];
+    const image = await createImageWithText(imageWidth, imageHeight, 50, 128, 'Hello World!');
 
+    // For debugging purposes, write the image to disk.
     image.write(__dirname + '/image2.png');
 
-    // Rotate image for label writer.
+    // Rotate image for label writer. Needs to be in portrait mode for printing.
     image.rotate(-90, true);
-    const bitmap = await imageServices.convertImageToBitmap(image);
+    const bitmap = await convertImageToBitmap(image);
 
     await dymoServices.print(bitmap);
     console.log('Successfully printed');
+
 })();